Corrected MACOSX installation

+++ DIFFUSE_INSTALL/install_jre_jmol.sh
@@ -7,7 +7,9 @@
 #
 export JRE_DONE=0
 export JRE_INST=0
-if [ -e /usr/bin/java ]; then
+DD=$(/usr/libexec/java_home)
+if [[ "${DD}" == "/Library/Java/JavaVirtualMachines/adoptopenjdk-11.jre/Contents/Home" ]]; then
+  echo "JAVA ${DD}"
 #
 #  Java exists 
 #
@@ -25,17 +27,18 @@ else
 fi
 
 if [[ $JMOL_INST == 1 ]]; then
-  curl -o OpenJava.pkg -fSL https://github.com/AdoptOpenJDK/openjdk11-binaries/releases/download/jdk-11.0.5+10/OpenJDK11U-jre_x64_mac_hotspot_11.0.5_10.pkg 
+  curl -o OpenJava.pkg -fSL https://github.com/AdoptOpenJDK/openjdk11-binaries/releases/download/jdk-11.0.7+10/OpenJDK11U-jre_x64_mac_hotspot_11.0.7_10.pkg 
   echo
   sudo installer -pkg OpenJava.pkg -target /
   export JRE_DONE=1
 fi
 #
 if [[ $JRE_DONE == 1 ]]; then
- curl -o JmolLatest.zip -fSL https://sourceforge.net/projects/jmol/files/latest/download
+  if [ ! -e JmolLatest.zip ]; then
+    curl -o JmolLatest.zip -fSL https://sourceforge.net/projects/jmol/files/latest/download
+  fi
   unzip -oq JmolLatest.zip
   export VERSION=$(ls | grep "jmol-")
-  echo $VERSION
   rm -rf $HOME/Downloads/$VERSION
   cp -r $VERSION $HOME/Downloads
   rm -rf $VERSION
